+++ DatadogSessionReplay/Sources/SessionReplay+objc.swift
@@ -124,9 +124,9 @@ public final class objc_SessionReplayConfiguration: NSObject {
     ///
     /// - Parameters:
     ///   - replaySampleRate: The sampling rate for Session Replay. It is applied in addition to the RUM session sample rate.
-    ///   - textAndInputPrivacyLevel: The way texts and inputs (e.g. label, textfield, checkbox) should be masked. Default: `.maskAll`.
-    ///   - imagePrivacyLevel: Image recording privacy level. Default: `.maskAll`.
-    ///   - touchPrivacyLevel: The way user touches (e.g. tap) should be masked. Default: `.hide`.
+    ///   - textAndInputPrivacyLevel: The way texts and inputs (e.g. label, textfield, checkbox) should be masked.
+    ///   - imagePrivacyLevel: Image recording privacy level.
+    ///   - touchPrivacyLevel: The way user touches (e.g. tap) should be masked.
     ///   - featureFlags: Experimental feature flags.
     @objc
     public required init(
@@ -134,7 +134,7 @@ public final class objc_SessionReplayConfiguration: NSObject {
         textAndInputPrivacyLevel: objc_TextAndInputPrivacyLevel,
         imagePrivacyLevel: objc_ImagePrivacyLevel,
         touchPrivacyLevel: objc_TouchPrivacyLevel,
-        featureFlags: [String: Bool]? = nil
+        featureFlags: [String: Bool]?
     ) {
         _swift = SessionReplay.Configuration(
             replaySampleRate: replaySampleRate,
@@ -144,6 +144,29 @@ public final class objc_SessionReplayConfiguration: NSObject {
             featureFlags: featureFlags?.dd.featureFlags ?? .defaults
         )
         super.init()
+    }
+
+    /// Creates Session Replay configuration.
+    ///
+    /// - Parameters:
+    ///   - replaySampleRate: The sampling rate for Session Replay. It is applied in addition to the RUM session sample rate.
+    ///   - textAndInputPrivacyLevel: The way texts and inputs (e.g. label, textfield, checkbox) should be masked.
+    ///   - imagePrivacyLevel: Image recording privacy level
+    ///   - touchPrivacyLevel: The way user touches (e.g. tap) should be masked.
+    @objc
+    public convenience init(
+        replaySampleRate: Float,
+        textAndInputPrivacyLevel: objc_TextAndInputPrivacyLevel,
+        imagePrivacyLevel: objc_ImagePrivacyLevel,
+        touchPrivacyLevel: objc_TouchPrivacyLevel
+    ) {
+        self.init(
+            replaySampleRate: replaySampleRate,
+            textAndInputPrivacyLevel: textAndInputPrivacyLevel,
+            imagePrivacyLevel: imagePrivacyLevel,
+            touchPrivacyLevel: touchPrivacyLevel,
+            featureFlags: nil
+        )
     }
 
     /// Creates Session Replay configuration.